+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 2023.01.15 v1.0.0
 
-1. create project
+1. add `utils` `lerna-workspace-root` `pnpm-workspace-root` `yarn-workspace-root` `workspace-root` `which-pm` `what-pm`